Controller for LinkField
part of: story #24968 [modal] have the list of linked artifacts

No functional change expected in new artifact link field.

Notes:
Using a Controller lets us separate the retrieval of the links from
their presentation. The hybrids component is now limited to
converting a Presenter to some HTML. Dealing with the state is delegated
to the controller: it returns a presenter for each case. The presenter
can even be immutable.

Change-Id: Ia68e1769e69a1730e48ed66fdfa61295947eb542