fix(webdav): request #25321 WebDAV support broken at PROPFIND
Original implementation was relying on extend of Dav\Directory & File
that supposed to be a wrapper for a filesystem exposed via WebDav.

However in our case it's not an actuall filesystem but an abstraction so
we should actually implement the interfaces instead of extending.

This solves the issue raised in the request because:
- we no longer implements IQuota so it won't trigger the issue
- the code for the generic method were expecting a $path to be
  initialized but as 100% of our extends were not calling parent
  constructor the code was broken (issue visible with PHP 8.0 argument
  check for internal methods).

Change-Id: I26b841c6927aed0813270bea651675c9ae2d2768