feat: URL based feature flag
Instead of relying on the platform administrator to activate the feature
flag, we now use an anchor in the URL of documents to activate the new
search. This offers two advantages:
* It is not anymore the-whole-platform-or-nothing,
* Since each user can activate it, we can deliver the feature sooner to
early birds.

If #feature-flag-new-search is present in the URL, then we do not search
anymore in the legacy docman. Since the feature flag is read during the
initialisation of the application, user can navigate in folders without
losing it (even it is not anymore in the URL).

Part of story #24197: search for Documents with criteria

Change-Id: I7b0eb2dbb3045a205120c1d541c80a8ee3f31eb0