Harden identifier typing
part of story #24969 [modal] add and remove links without types

No functional change expected in the "Parent feedback" and the new
Artifact link field

Notes:
This lets us stop primitive obsession with artifact ids. It even
simplifies some of the code because we can deduce from the identifier if
the modal is in creation or edition mode.

Change-Id: I61b3353ed0f5140b5bdf41bb865d6e2f72776186