"Static method" style
part of story #24969 [modal] add and remove links without types

No functional change expected in the "Parent feedback" and the new
Artifact link field

Notes:
In PHP we use static class methods to build objects. It offers good
readability: we can build a Presenter "from" an Artifact, for example.
Same thing for stubs: we can build a stub "with" a return value.

Since we are not using classes in TypeScript, we used simple factory
functions. However, it is harder to name them to have the same
readability without writing very long names. We can "simulate" static
class methods by exporting a const object with the name we want.

We could also use "import * as Name" pattern, but this naming cannot be
guaranteed to be consistent, as any class is free to rename its imports.

Change-Id: I210cf916cc0070c4343fd1d9f00bc05bae2270bc