feat: Item results should have link
What's the point to be able to search for documents if we cannot
interact with them?

This contribution adds a link to every items (except empty) in order to
access its content:
* Follow a Link
* Download a File
* See an Embedded
* Access to a Wiki
* See the content of a Folder

Note: do not forget the feature flag to test the feature.

Part of story #24197: search for Documents with criteria

Change-Id: Ibc256f7a0bdfdba1485199e6751682865c557816