Using Faults in artifact-modal
part of story #24969 [modal] add and remove links without types

No functional change expected when fetching a parent artifact or
fetching linked artifacts.

Notes:
Using Faults lets us move more logic to the Domain instead of the
Controllers. The Domain now knows that if there is no
ParentArtifactIdentifier, it means we are in creation mode. It tells the
Controller (via a Fault) to create an empty presenter. The Domain still
has no knowledge of Presenters, but it handles the different cases,
instead of letting the Controller decide this.

Change-Id: I84c30ab7aa02a5041aa2a162831f72be2fd861a3