feat: search in whole project documentation
When we search in the whole project documentation (aka in root folder),
then no breadcrumbs should be displayed in the search criteria panel.

When we search in a sub folder, then an additional link "Search in whole
project documentation" is added after the breadcrumbs.

Part of story #24197: search for Documents with criteria

Change-Id: Ie5d4d10dca6b927135be729030178af0c6e33ca8