Initialize REST endpoint `GET /projects/:id/3rd_party_integration_data`
This endpoint will be needed for independant services to have access to
the information required by the Tuleap components. Those independant
services will be able to access this endpoint using an OAuth2 access
token with `read:project` scope.

This contribution sets the grounds for this new REST endpoint but it is
not yet complete. The sidebar `config` attribute will come in another
contribution.

Part of story #24227: have private oauth2 endpoints for Mediawiki

Change-Id: I47d4860709d3c51a1c482b982dc4226eb94a2aa2