Make possible to store generated front-end assets under each package root
Currently every front-end assets needs to go under src/www/assets/,
while this work it requires additional management when packaging the
sources and it is also challenging to cache the build result. In the JS
ecosystem it is expected that the build results of a package are located
under its root and not elsewhere. Typically, turborepo expect stable
outputs directories [0] for each task you want to cache, with our
current situation this is not something we can provide.

This contribution make posible to store the front-end assets inside a
`frontend-assets` in each package root (so in your case in each plugin
since we have a package for each plugin) in addition to storing them
under src/www/assets/.
Only the API Explorer and the DocGen plugins are affected by this
change. Others plugins (and core) will be managed in other
contributions. In situation requiring it (AFAIK only the tracker plugin
is concerned) an additional `backend-assets` will be created in order to
store the backend files that are not generated through Composer.

To test, do not forget to redeploy nginx configuration. No functionnal
changes is expected.

Part of request #24241: Cache the build results of the JS toolchain

[0] https://turborepo.org/docs/features/caching

Change-Id: Ieb620fc523e9218d8be92660389b1f4853417049