Make possible to remove the collapse button of the project sidebar component
The button can be removed using the `no-collapse-button` on the HTML
element.
When the button is removed the sidebar state is only managed by the
presence (or absence) of the `collapsed` attribute.

The easiest way to test the change is to go in one of the sidebar
package root and to start a dev server with `pnpm run dev` and to play
with the element.

Part of request #23465 Publish a project sidebar custom element

Change-Id: Ia1ce684a6256f0d12db8d821cc2856804142c4e1