Toggle removal of link in UI
part of story #24969 [modal] add and remove links without types

How to test:
There is a button at the right of the links table that allows to mark a
link for removal. When clicked, it will add a "strike-through" effect on
the link's row.

When clicked again, it will remove the "strike-through" effect.

Marking a link for removal will not actually remove it yet, it only
affects display for now.

Change-Id: Ie6007848626d0081f57a0cbc640446448f39118e