Add @tuleap/tlp dependency
part of request #24241 Cache the build results of the JS toolchain

This dependency does not change anything at runtime but is needed for
typechecking. Our tsconfig.json configuration aliases "import from tlp"
to the actual source code in src/themes/tlp/. The TLP package in turn
depends on @tlp/popovers. If we don't define this @tuleap/tlp
dependency, our build tools won't build @tlp/popovers. If it is not
built, tsc cannot find the types for @tlp/popovers, which will result
in a (seemingly) unrelated error like:

  @tuleap/plugin-mytuleap_contact_support:typecheck: ../../src/themes/tlp/src/js/index.ts(26,15): error TS2307: Cannot find module '@tuleap/tlp-popovers' or its corresponding type declarations.

Change-Id: Icdc9c35cfed1917f744ecb2d5c3a21402e90df9b