chore: Fix (again) documents e2e tests
In order to improve:
* readability
* separation of concerns
* repeatability

search and phpwiki tests have now their dedicated spec files.

It is really a PITA to debug existing tests when each tests depends on
each others. Inclusion of tests for search[0] should not have any
incidence of existing phpwiki tests. We should be able to run multiple
times a spec file with `make tests-e2e-dev && npx cypress open`.
It is now the case for phpwiki.spec and search.spec, document.spec will
be done later when time will permit it.

Part of story #24197: search for Documents with criteria

[0]: git #tuleap/stable/57345a3133b0526f69ec4e38aedd3c242993d392

Change-Id: I0f1d38029dbfc00db7e4f9ad414c82664e5c451b