chore(forge-config): Use ForgeConfig to carry database configuration
DBSetupParameters was introduce to ease the management of the
configuration at setup time but I realized it was a bad move related to
generalization of the approach.

In order to manage local.inc, I will need to re-use the code that sits
now in DatabaseConfigurator (getVariableForConfigFile) but this code was
managing explicitly the value to serialize and, worst, default values
were managed explicitly while now we can define those default values
directly on the ConfigKey.

Part-of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: I844c5716b968507a54f78aa4635318a2b8d813f1