chore: add app_type column in plugin_oauth2_server_app SQL table
This is part of story #24227: have private oauth2 endpoints for Mediawiki

Mediawiki will be able to define its own OAuth2 apps.
Before moving this SQL table to the core, we add the app_type column
that will be able to destinguish between MW OAuth2 apps and other apps
defined by the site administrators.

Change-Id: I1068c19462122c2440ac919c6d4fe78b5508a9d5