chore(forge-config) Write local.inc with TuleapCfg
This is part of the work to transfer most of the configuration to Tuleap
code so Containers will have access to a flexible installation while the
code stays as close as possible to the non container install.

This commit propose a new approach for local inc: only set the values
that are rewritten. As we load `local.inc.dist` behind the scene it
should give a better understand of what was overloaded by local lord.

In order to leverage the work started on database.inc serialization, the
code is transfered and re-used. At this stage, the variables in ServerHostname
and ConfigurationVariables classes is only there for local.inc
serialization. The work to have those variables properly managed
end-to-end will come later on.

Part of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: Ia1f710894a2978e8d914e2845b365d4b7aa14d4c