Convert file upload quota display to hybrids
Part of request #22648 : Convert AngularJS directives and controllers
to hybrids web components

How to test:
- No functional change expected in the display of the user's temporary
file upload quota. If no usage, it shows the available space. If there
is usage (you can fake it with REST or by editing rest-service.ts), it
shows how much is being used with a progress bar.

Notes:
I intend to replace "file-upload-rules-state" by
UserTemporaryFileQuotaStore. This is an intermediary state where both
exist to reduce the size of the change.

Change-Id: I5629e15f3f00f7353c322befb92c05f9688b72d8