chore(forge-config): Transfer dbauth user & password to database
The amount of things changed is, unfortunatly, high because of the very
nature of those variables:
- used heavily at the early stage of the application life (installation)
- a password that should be stored encrypted but that wasn't

Notes:
- SetupMysqlInitCommand cannot re-use ConfigSet code to store the
password because tuleap-cfg doesn't load pre.php so the database is not
connected. So I had to write down the SQL queries directly. That's not
ideal but I think the tradeof is acceptable at this point.
- SetupMysqlInitCommand need a dirty workaround (ForgeConfig::set)
  because there is a dependency on it in KeyFactory.

This step is mandatory to be able to auto generate local.inc.

Part of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: Iecac5f8199474a64ca23b9989be46c0314469f1d