request #25354 Double list picker on list field
How to test:

1. Make sure list-picker feature flag is active
2. Create an artifact with a list field set to "None"
3. Make the list field "required" in tracker admin
4. Set up a field dependency so the list field is "target"

When you edit the artifact, the list field should be "in edition"
because it's required and empty. When you click to edit the "source" of
the field dependency, the list field should only have one list-picker.
Without this patch, there were two list-pickers for the same field.

Change-Id: Iacd633bbafd9833b21b8c2181f4a4d1f96e4b51a