Project sidebar is now a block element with an explicit height and width
By default custom element are considered to be inline element which does
not make a lot of sense for our situation. Also we know set a width and
height to the component so it is easier to do manipulation on it (e.g.
moving of XXXpx from the top to add a platform banner).

Part of story #24227: have private oauth2 endpoints for Mediawiki

Change-Id: I3cfa4d113a9ab6e60a4abe313e4766897e0f30ea