fix: Docman admin makeover broke e2e tests
Link to create a new document was removed. Therefore we have to go back
to the project documentation root before. Breadcrumbs template is updated
to be able to add data-test attributes on them (only subitems were
supported before).

The docman e2e has been updated to use data-test attribute instead of
label to create a new document, and updated also to be able to run
multiple times the tests in the Cypress Test Runner.

Part of story #24197: search for Documents with criteria

Change-Id: Ie25c013f6079eacd70a1ba19419d39e9dfe4fe65