feat: Ability to search against custom properties
If project has defined custom properties, then corresponding filter
should be displayed in the search criteria panel.

No functional changes expected for existing hardcoded properties.

Note: it is not possible to offer the same REST API UX between type and list
fields. Type (as well as status) values are hardcoded and can be used
directly in the REST API ("folder", "wiki", …). However since it is
possible to have multiple values with the same label, we can only use
their internal id (102, 103, …).

Part of story #24197: search for Documents with criteria

Change-Id: I7ffe51577b2a1f1b24b656323059591e4dd3eb32