Make sure popovers of the project sidebars are not hidden by other elements
The issue is particularly visible on the planning view of the Agile
Dashboard.

When removing the specific implementation of the FP and BP project
sidebars I thought it was a whole new world and we did not need the
z-index hacks, I was wrong.

Part of story #24227: have private oauth2 endpoints for Mediawiki

Change-Id: Ibdf669e0e2868c18dd05f1b2592324ea422a22d6