refacor: Rearrange Search route parameters
For better consistency of searching for items via REST API, the
docman_search/:id route is rewritten to accept the following format:

```
{
   "global_search": "lorem*",
   "properties": [
     {"name": "status", "value": "draft"},
     {"name": "field_3", "value": "lorem*"},
     {"name": "create_date", "value_date": { "date": "2022-01-30", operator: ">" } },
   ],
   "limit": 50,
   "offset": 150
}
```

The added bonus is that it is now easier to define in TypeScript the
expected format for the request body (no moar soup of Record<>).

No functional changes is expected.

Part of story #24197: search for Documents with criteria

Change-Id: Ia49275574ef11811d740ca4701fd24378c45e7eb