Create @tuleap/vue3-gettext-init lib
This library will be used by Vue 3 apps to init their Gettext provider.

For now this library is not directly used since we need an app using
Vue 3 and Gettext. Some changes might be made to it once we start really
using it.

Part of request #25365: Migrate the TestPlan plugin to Vue 3

Change-Id: I72abadeb3783275c8a8c5fdb30b3ed44d299811d