feat: Convert docman perms admin to BurningParrot
In order to bring some consistency to admin pages (like Obsolete or
Locked documents) the document » administration » permissions page is
converted to BurningParrot.

The difficulty here is that we depend on legacy functions like
permission_display_selection_form() that directly output a form with
needed inputs. While this is good to have a common lookalike form
everywhere, it is hard to style once we display it under BurningParrot.
The solution brought here is to add some javascript to add css class on
components that we cannot control directly. This avoid modifying
existing code and introducing #regressions.

Part of story #24197: search for Documents with criteria

Change-Id: I16b00f43d84b1313c49ebf18ff1656543651f39e