Convert File field to hybrids
part of request #22648 Convert AngularJS directives and controllers to
hybrids web components

How to test:
No functional change expected in the file fields. You can still attach
multiple new files (with description). On submit, the files will be
uploaded and attached to the artifact. There is a "Reset" button that
lets you clear the attached file and its description.
For existing files, you can mark them for deletion. Unless you cancel
that deletion, on submit the files will be removed from the artifact.
Please check that the "required" behaviour is still active and that the
field is disabled if you don't have permission to update / submit the
field.

You can test basic non-regression with cypress end-to-end tests for
agiledashboard.

Notes:
- angular-sanitize could probably have been removed earlier, I just
noticed it now.
- It would be better from a hexagonal architecture point of view to
handle the "Files marked for removal" and generally the field data in a
Memory adapter, however it produced a change too big to review so this
patch focuses on the template (which is the more pressing issue). The
controller mutates the "value model" for now, it can be refactored
later.

Change-Id: Iccfe9eb2b8dcbc2fffea02b254e51fba27ee1429