fix: Redirect the user when trying to access to docman administration
Since admin page is displayed with BurningParrot, whenever a project
member tries to access to administration pages, the page displayed is
screwed up: the feedback error message is displayed and the docman
content is displayed. However there is no redirection and the styles are
wrong: legacy docman (FP) with BP theme.

This commit add a redirection to fix that.

Part of story #24197: search for Documents with criteria

Change-Id: I31dd4828b6d2db3046fb9f0f1c3070f808f2de39