feat: Display obsolescence date column in search results
As soon as the obsolescence date property is used, the corresponding
column is displayed in search results.

If an item has no obsolescence date defined, nothing is displayed for
now. A design check will probably change thinks when we will have more
columns to better see how it fits.

Note: CellString has been updated to use a slot instead of a prop, for a
simpler usage.

Note: Existing Cell* tests have been updated to have only needed Item
properties for the test. This will prevent to update every tests
whenever we add a new property in the Item interface.

Part of story #24197: search for Documents with criteria

Change-Id: Ica2ad76521db6eeaaf5fd726c1f9745319811684