Provide a specific OAuth2 authorization endpoint for the MediaWiki standalone plugin
The only differences with the OAuth2 server endpoint is that only a few
scopes can be requested but they are always auto approved without
consent from the users. This situation is acceptable because we trust
the MediaWiki installation.

The only allowed scopes are the ones needed for the integration to work
so information about the users, the sign-in scope to authenticate the
users and read-only informations about projects.

The generic OAuth2 server authorization endpoint cannot be used to
deliver an auth code for the specific "MediaWiki" apps and vice versa.

Functionnal tests is a bit tedious at this stage since there is nothing
generating a "MediaWiki app" (so no client ID/client secret). However
the generic OAuth2 server authorization endpoint should continue to work
as before (you can check that using the E2E tests of the OAuth2 plugin).

Part of story #24227: have private oauth2 endpoints for Mediawiki

Change-Id: I4653f9b9c07fec9c1662c32fc51cf8996572cdc1