feat: display custom properties as column
If docman administrators defined custom properties (text, string, date,
list, multiple list), then the should appear in the search result
columns and the corresponding item's value is displayed.

For list with multiple values, they are displayed as a comma separated
list of values, maybe we should find and alternative? A dedicate
contribution will handle this if needed.

Part of story #24197: search for Documents with criteria

Change-Id: I01a7ea5005b29549db27db21e870aef3c2a99806