feat: Move filename pattern admin in extra tabs
The feature of patterns in filename is considered less important than
managing properties, or permissions, and we don't want to promote too
much it. Therefore we "hide" it under the hamburger menu, like for
legacy docman preference default view.

Part of story #24166: apply a naming pattern on files to enforce a convention

Change-Id: Ib07a6552f80e3b3e9ba5600cca08af42adcdf44c