chore: Convert Property administration to BurningParrot
Follow-up to b4cabd337839f99c6ed8d8494f93ba806685cd51

➡️ Page is displayed under burning parrot theme
➡️ Modals are used to create value in list properties
➡️ Popovers are used to confirm deletion of values in list properties
➡️ CSRF token are added to guard update of property, and creation or deletion of values

Note: I'm not really satisfied with the back link (to return to the list
of properties, or to return to the property itself when we edit a
value). However I think that current proposal is better than current
situation so we will see later with the Design team how we can improve this.

Note: Edition of a property value could be done now in a modal instead
of a dedicated page. Same, to be checked with the Design team.

Part of story #24197: search for Documents with criteria

Change-Id: I528c0f51c5372c7713682d27595c2c2853652559