Closes request #25678: Git administration page is broken
Issue was introduced by git #tuleap/stable/b9fb3c11bfd1d8b608a3f577c03cabb197a6612d

This contribution makes sure we determine if we are in a BP section of
the document manager administration when we display the document manager
admin page. Until now the verification was not scoped so any pages with
a parameter `action` set to `admin` was considered a BP page.

The event Tuleap\BurningParrotCompatiblePageEvent is deprecated to
discourage its usage. It is expected that determining if a page is
BurningParrot page or not to be scoped to a specific route. The way the
event is designed you either do not check that (leading to the current
issue) or you are duplicating what the router already does.

Change-Id: Ib42faaaf11dab8703fde31394f478137a089e1a2