Add an additional button in the report export menu for the XLSX export
The button is only visible when a feature flag is enabled with
`tuleap config-set feature_flag_doc_gen_cross_tracker_report_xslx_export 1`.

For now clicking on the button does nothing (well it throws an error
message in your web browser console).

A small changes has been made in the way tracker report are rendered so
we can properly import Vite assets (with their stylesheets). Now the
tracker header is outputted only once we collected the assets we need on
the page.

Part of story #23469: export reports with linked artifact in xlsx format

Change-Id: I54f228eaa138fdd9bc44eff33654c8ecaf2fc14b