xlsx: 0.17.5 -> 0.18.3
The same lib is going to be needed to generate the cross report link
based document.

Changelog:
https://github.com/SheetJS/sheetjs/blob/v0.18.3/CHANGELOG.md
https://github.com/SheetJS/sheetjs/compare/0ddeba4fc358394438e3651910bce3b3ef29578a...e14aee3e512a838a9d2a125568c8ae8070695957

They have cleaned up a bit the way the library is built and what's
embedded which result in a smaller payload for our usage.

Part of story #23469: export reports with linked artifact in xlsx format

Change-Id: I3089e09f4161a4b93c1dd13f3bde0aec5adf5f54