feat: Add option to forbid Writers to update item
This option is hidden under a temporary feature flag (that should be
removed in the next commit): add feature-flag-forbid-writers=1 in the
docman » admin » permissions page url.

The settings is saved, but not taken into effect yet.

CSRF has been added to the form to secure it. Due to the fact that the
form is still managed by legacy code, we had to cheat a little bit and
move dom elements with some javascript. => to avoid flickering the form
is hidden until the CSRF and other addendum have been inserted.

Note: [reset to default] button has been removed. It prevents us to
properly add the checkbox in the form (or we will have to bind the
checkbox behavior to this button). Furthermore, except in PhpWiki, we
don't have such buttons in permissions form in Tuleap. So, with Design
Team's approval, we don't have anymore the reset button on docman
permissions.

Starts story #24980: forbid modifications of document properties by WRITERS

Change-Id: If6b30a03f7e88435e02ae0b573cde96e0ce0186d