feat: Add from_table_renderer parameter in GET /tracker_reports/{id}/artifacts
This is part of story #23469: export reports with linked artifact in xlsx format

A new parameter named from_table_renderer is added to the GET
/tracker_reports/{id}/artifacts REST route.

If provided, this parameter will allow to return the matching artifacts
with the fields used in the table renderer of the report. Artifact link
field is not returned in these representations, because of the multiple
columns (with or without specific types) that can exist (and because this
field will be used in a different way in the XLSX document).

If the report has no table renderer or mulitple ones, then an error will
be raised.

Change-Id: Ifae866e5f4090471c95bb0ea72ea490c1b036071