Correct tracker artifact e2e tests
part of part of request #24982 clean e2e test suite

It might occur that connected user has more than a dashboard tab.
The test was assuming that user only have one.

I didn't manage to reproduce the issue in dev mode, and I don't reproduce the issue in every full run.
I'm assuming that it is due to a test who fail at first run and create a new tab in `my personnal page`

Note:
the line `cy.get("[data-test=add-to-my-dashboard]").first().click({ force: true });`
is crappy, but we choose to do this because the other solution was to define a dynamic data-test
on personal page tab name combinated to a if in tracker report view to check if user has a button or a dropdown

the `.first` isn't an issue as once the report is added, user is redirected on personnal dashboard for doig assertions
the `force:true` is not for best but it prevents us to have to check if we have a button or a dropdown:

Change-Id: Ie22cc8b0d8a6bf33d352252243473a6f6768ef2f