story #24981: forbid deletion of documents and versions by WRITERS
The behaviour is configured in "Manage permissions" page of docman administration

there is a switch "Forbid deletion of document & versions by Writers".
When activated, only Managers or Docman Administrators can delete stuff
in docman/documents.

This applies:
* in SOAP API
* in REST API
* in WebDav
* in docman "legacy" interface
* in Document app

Note: \WebDAVDocmanFile::delete() is updated so that it return
MethodNotAllowed the same way \WebDAVDocmanFile::rename() does.

Change-Id: Icde65e5874b6821d0e8590adb1060224d240c439