Migrate TestPlan app to Vue 3 in compat mode
Everything should continue to work as expected.
There a few warnings when the app is built in dev mode, they will be
solved in dedicated contributions.

Typechecking has been switched over to vue-tsc because it was the
easiest thing to work with (and its hability to typecheck templates is
useful in the migration process).

Part of request #25365: Migrate the TestPlan plugin to Vue 3

Change-Id: Ib2be8520b05c71935b9b4da12fca0e611ee9ab38