feat(forge-config): Set admin email in the docker image
Re-use $sys_email_admin for default admin email:
- site administrator account email
- default admin email
- catch all email alias

There was `TULEAP_EMAIL_ADMIN` variable that can define
the forward email for `codendi-admin@$domain-name` (postfix aliases), I
removed this variable in favor of TULEAP_SYS_EMAIL_ADMIN as it
semanticly targets the same person.

This commit changes the way we deal with Configuration during the
platform init. Until now we assumed that \ForgeConfig was more or less
empty and loaded locally the variables when needed (with store/restore
or wrap). There is actually too many usage of ForgeConfig to safely
assume that it cannot be init. So Tuleap class will load the
configuration, whether it's at init or at restore time.

Closes request #21742 Set admin email in the docker image

Part of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: I98dd2940d0747898142a4fb6c13816edb47eac9b