feat: complete support of internationalized characters
As of today, file names are stripped at download to avoid clash on non
ISO-8859-1 chars. In order to support titles and status in documents
download when a filename pattern is set, we should complete the support
of internationalized characters as per [RFC6266][1].

The generated filename may contain `/` or `\` which are unsupported
characters (but legit in a title) we replace them with dash.

In order to test this, I suggest to activate the filename pattern in
document[2] and then unleash your creativity to create File document with
title containing emojis, kanji, cyrillic, … Go crazy!

Bonus: bump of vfsstream in xml-template-to-php to be able to run tests
in PHP8.1

Part of story #24166: apply a naming pattern on files to enforce a convention

[1]: https://www.rfc-editor.org/rfc/rfc6266#section-6
[2]: `tuleap config-set feature_flag_apply_naming_pattern 1`

Change-Id: Ia604f0a88eaec91255365441c113fddb47aaca53