Fix gettext extraction if the cross report document export modal
The build manifest was not properly configured initially so the
extarction was not correct. This contribution adds a bit more of content
to ensure it works as expected in different contexts.

The added content does not deal with the styles nor the behavior. For
now it is purely static.

Part of story #23469: export reports with linked artifact in xlsx format

Change-Id: Ic49f7876f1d24b580eadbecef37e3ae239fb72c2