feat: Disable webdav if filename pattern is set
As soon as a filename pattern is set, we cannot use anymore WebDAV for
the project Documents.
(sabredav seems to not handle very well the on the fly renaming).

If webdav is used by (and allowed for) the project, then a warning is
displayed in filename pattern administration. Took the opportunity to
redesign a little bit this page.

Note: Update of webdav tests is to make sure we can run them in PHP8.1

Part of story #24166: apply a naming pattern on files to enforce a convention

Change-Id: I30578b2f35b182f06997b965f509d40c3c483ed4