fix(forge-config): setup local.inc can no longer rely on sed
Since 563b27155afe1c6f000b43574094c103f63cd05a local.inc only contains
the variables that must be overriden at install. Existing deployments
that were relying on sed of default values (because previously the whole
local.inc.dist was copied) no longer works.

This commit re-align those deployments with the new practices and takes
the opportunity to align the integration tests (rest, soap, etc) with
standard installation tools so the tests are as close as possible to a
real installation.

For distlp, the management of trusted proxy was entirely removed because
it's no longer needed (as the tests continued to pass despite the sed
that was not doing anything).

Part of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: If2150759a7d8b2fd5179cb7e688d1e7a601cc366