Extract translations from TestPlan SFC as Vue 3
The translations are now extracted using the Vue 3 process (which make
possible to use Vue 3 specificities in SFCs without issues).

`vue3-gettext` deprecates the usages of the `<translate>` component and
of the `v-translate` directive and it's PITA to extract from the
templates so we do not even bother. The existing usages are moved to the
equivalent translation functions.

No functional changes is expected from this contribution. Note that
displaying the content in French is currently broken but this unrelated
to this change and will be fixed in a dedicated contribution.

Part of request #25365: Migrate the TestPlan plugin to Vue 3

Change-Id: Id13c7e631148de4021982667e9bc2f4accf9e17d