Fix display in French of the TestPlan app
The issue was introduced by the switch to vue3-gettext in
98430abdf46f0842743d8e0111df367e0438c11f.
The @tuleap/vue3-gettext-init library expects to have a file transformed
with gettext-parser instead of easygettext to make the cohabition with
our other gettext usages easier (same transformation can now be done
everywhere).

To do that we now provide a plugin that can be used either with Vite or
Webpack to replace our easygettext rules and transformations.

Part of request #25365: Migrate the TestPlan plugin to Vue 3

Change-Id: Ib96557bd447c96a11d7b1a78e44fc51cafbdc527