Temporarily allow to retrieve export properties from URL parameters
This contribution make possible to use URL paramaters to export content.

For example if you add the GET parameter `export_report_id_1=135` to
your URL, the report with the ID 135 will be used. Same approach can be
used to select level 2 and level 3 reports as well as the expected type
of links to use for each level.

Ultimately this will be replaced by a web UI in the modal.

Part of story #23469: export reports with linked artifact in xlsx format

Change-Id: Ic7aac09e76b9b6d03caf04ad9a1b952c7284b44a