feat: Checkbox to enforce filename pattern
Now siteadmin better knows how to deactivate the feature thanks to a
dedicated checkbox to enforce (or not) the filename pattern.

The admin cannot enforce an empty pattern.
The admin can keep the pattern but disable the enforcement.

Part of story #24166: apply a naming pattern on files to enforce a convention

Change-Id: Ib68934f57e990ce444631746067b2ff03dcbead1