feat: Pimp my filename pattern administration
Add some explanations and example about the feature so that the user
better knows what is it used for.

Display variables in a table for better readability.

Do not assault the admin with a red text about mandatory variables when
they doesn't enter anything yet.

Add title attribute to the text field for a more explicit message when
the pattern of the pattern (Yo dawg) is not respected.

Decrease the level of WebDAV message if pattern is not enforced.

Part of story #24166: apply a naming pattern on files to enforce a convention

Change-Id: I8a083013642d7b68ed3b6a4e98bd498f46b2818f