refactor: stop using Tracker_ArtifactLinkInfo
part of story #24969 [modal] add and remove links without types

No functional change expected.

Tracker_ArtifactLinkInfo can call the database and will make us mock it
all the time. We should use our own concept of link instead.
I also put an interface in front of ArtifactLink because it relied too
much on REST payload format. We are not doing hexagonal architecture
here, but it would still be nice to have a separation between "sort-of
domain objects" and the REST API.

Change-Id: Icbea62a4228f09d5c09e30fec57e230a2f633cd4