feat: Display dropdown for search results
Description
===========

Every search result items now offers a dropdown menu in order to be able
to fully interact with it. Every actions that are available when we
browse the tree are available for each type of items.

When an item is updated and modification can have an impact on displayed
search results (update of properties, new item, permissions, …) then the
whole page is reloaded.

For now the label of the button is only an ellipsis, we might change it
later to a text label if we find inspiration.

Implementation notes
====================

Delayed display of dropdown
---------------------------

Since the result is an array of ItemSearchResult, which is not
compatible with Item, we need to do a query to retrieve the full item
when we open the dropdown.

Blocking modal
--------------

When a new version (or file item) is being uploaded, we don't want the
user to screw up the upload with navigation. The user should be aware
that an upload is ongoing. While the current folder has progress bars
(in header and in content) it is not the case yet in search result.

The easiest/fastest solution is to display a blocking modal until the
upload is finished. So now the TLP modals accept a new option in order
to control the click on the backdrop (click to close), to be used in
combination with `keyboard` option (ESC to close).

Externalized dropdown menu
--------------------------

The structure of the search results DOM (horizontally scrollable table)
prevents the display of the dropdown menu. When it is displayed,
according to situations/resolutions it might be hidden behind the
overflow, forcing the user to scroll the table to see menu options.

However TLP dropdowns must be initialized with menu as a sibling of its
trigger button. In order to solve the issue, once the TLP dropdown is
initialized for search results items, the menu is moved directly at the
end of `document.body`. That way the floating menu is displayed above
all content and is not anymore under the overflow of the table.

Manipulating DOM element inside Vue component sounds a little bit
cheaty, but I didn't find a better way to achieve the result unless to
rework the WOW of TLP dropdowns.

---

Part of story #24197: search for Documents with criteria

Change-Id: I18b7a77280263770441cf7f3c52a372940f4ef83